+++ Example/PaymentSheet Example/PaymentSheetUITest/PaymentSheet+AddressTests.swift
@@ -48,7 +48,7 @@ class PaymentSheet_AddressTests: XCTestCase {
 
         app.textFields["State"].tap()
         app.pickerWheels.firstMatch.adjust(toPickerWheelValue: state)
-        app.toolbars.buttons["Done"].tap()
+        app.stp_dismissKeyboard()
 
         app.textFields["ZIP"].tap()
         app.typeText(zip)
@@ -326,7 +326,7 @@ US
         // Set country to New Zealand
         app.textFields["Country or region"].tap()
         app.pickerWheels.firstMatch.adjust(toPickerWheelValue: "🇳🇿 New Zealand")
-        app.toolbars.buttons["Done"].tap()
+        app.stp_dismissKeyboard()
 
         // Address line 1 field should not contain an autocomplete affordance b/c autocomplete doesn't support New Zealand
         XCTAssertFalse(app.buttons["autocomplete_affordance"].exists)
@@ -405,8 +405,7 @@ NZ
         app.typeText("San Francisco")
         app.textFields["State"].tap()
         app.pickerWheels.firstMatch.adjust(toPickerWheelValue: "California")
-        app.toolbars.buttons["Done"].tap()
-        app.typeText("California")
+        app.stp_dismissKeyboard()
         app.textFields["ZIP"].tap()
         app.typeText("94102")
         app.buttons["Save address"].tap()
@@ -424,7 +423,7 @@ NZ
         app.buttons["Address"].tap()
         app.textFields["Country or region"].waitForExistenceAndTap()
         app.pickerWheels.firstMatch.adjust(toPickerWheelValue: "🇺🇾 Uruguay")
-        app.toolbars.buttons["Done"].tap()
+        app.stp_dismissKeyboard()
         app.buttons["Save address"].tap()
 
         // ...should update PaymentSheet.FlowController
@@ -445,10 +444,10 @@ NZ
         app.buttons["Address"].tap()
         app.textFields["Country or region"].waitForExistenceAndTap()
         app.pickerWheels.firstMatch.adjust(toPickerWheelValue: "🇺🇸 United States")
-        app.toolbars.buttons["Done"].tap()
+        app.stp_dismissKeyboard()
         app.textFields["State"].waitForExistenceAndTap()
         app.pickerWheels.firstMatch.adjust(toPickerWheelValue: "California")
-        app.toolbars.buttons["Done"].tap()
+        app.stp_dismissKeyboard()
         app.buttons["Save address"].tap()
 
         // ...should not affect your billing address...
@@ -483,7 +482,7 @@ NZ
         // Select UK for phone number country
         app.textFields["United States +1"].tap()
         app.pickerWheels.firstMatch.adjust(toPickerWheelValue: "🇬🇧 United Kingdom +44")
-        app.toolbars.buttons["Done"].tap()
+        app.stp_dismissKeyboard()
 
         // Ensure UK is persisted as phone country after tapping done
         XCTAssert(app.textFields["United Kingdom +44"].exists)
@@ -575,7 +574,7 @@ NZ
 
         stateField.tap()
         app.pickerWheels.firstMatch.adjust(toPickerWheelValue: "New York")
-        app.toolbars.buttons["Done"].tap()
+        app.stp_dismissKeyboard()
 
         postalField.tap()
         let existingPostal = postalField.value as? String ?? ""