+++ StripeFinancialConnections/StripeFinancialConnections/Source/Helpers/FinancialConnectionsAppearance.swift
@@ -50,9 +50,9 @@ struct FinancialConnectionsAppearance: Equatable {
 
     init(
         theme: FinancialConnectionsSessionManifest.Theme?,
-        brand: LinkBrand? = nil
+        linkBrand: LinkBrand? = nil
     ) {
-        let resolvedBrand = Self.resolveBrand(theme: theme, brand: brand)
+        let resolvedBrand = Self.resolveLinkBrand(theme: theme, linkBrand: linkBrand)
 
         switch theme {
         case .linkLight:
@@ -74,31 +74,20 @@ struct FinancialConnectionsAppearance: Equatable {
         }
     }
 
-    private static func resolveBrand(
+    private static func resolveLinkBrand(
         theme: FinancialConnectionsSessionManifest.Theme?,
-        brand: LinkBrand?
+        linkBrand: LinkBrand?
     ) -> ResolvedBrand {
-        switch PresentationManager.shared.configuration.linkBrand {
-        case .link:
-            return .link
-        case .onelink:
-            return .onelink
-        case .unparsable, .none:
-            break
-        }
-
-        switch brand {
-        case .link:
-            return .link
-        case .onelink:
-            return .onelink
-        case .unparsable, .none:
-            switch theme {
-            case .linkLight:
+        switch theme {
+        case .linkLight:
+            switch PresentationManager.shared.resolvedLinkBrand(manifestLinkBrand: linkBrand) {
+            case .onelink:
+                return .onelink
+            case .link, .unparsable, .none:
                 return .link
-            case .light, .dashboardLight, .unparsable, .none:
-                return .stripe
             }
+        case .light, .dashboardLight, .unparsable, .none:
+            return .stripe
         }
     }