+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/Link/WalletViewControllerTests.swift
@@ -91,7 +91,7 @@ private extension WalletViewControllerTests {
                 intent: intent,
                 elementsSession: elementsSession,
                 configuration: configuration,
-                linkBrand: configuration.resolvedLinkBrand(elementsSession: elementsSession),
+                linkBrand: configuration.resolvedLinkBrand(elementsSession: elementsSession, linkAccount: nil),
                 shouldOfferApplePay: false,
                 shouldFinishOnClose: false,
                 initiallySelectedPaymentDetailsID: nil,